requirements.txt
numpy==1.18.1
pandas==1.0.3
matplotlib==2.2.5
django==3.1.4
numpy-stl==2.13.1